Fixed a bug where the demo would crash occasionally.
The palm detector sometimes returns a box frame of size 0x0,
which then throws an exception when the code tries to calculate the
average depth, as it gives the averaging function an empty list, which
then in turn returns a "NaN" object.
This commit should fix the problem, by checking if the box size is 0.